More Warning Fixes

+++ src/key.cpp
@@ -411,7 +411,7 @@ bool CKey::Verify(uint256 hash, const std::vector<unsigned char>& vchSigParam)
     {
         unsigned char nLengthBytes = vchSig[1] & 0x7f;
 
-    if (vchSig.size() < 2 + nLengthBytes)
+    if (vchSig.size() < (long long unsigned int)2 + nLengthBytes)
              return false;
 
         if (nLengthBytes > 4)
